Changed Stock class to store items in a HashMap
Changed Stock class to store items in a HashMap as later down the line, getting quantities for individual items from a list would require the system to go through the entire list of individual items, incrementing a final quantity value. Whereas with this, one value is stored per item and that can be returned with one command.