Made some changes, comments and a fix
Quite a few things done here. I didn't change actionPerformed like you told me to, but instead I implemented it in comments. Next, you don't need to declare a class as null; it's automatically set to null if no value is assigned to it. Also I commented out the List Items I was talking about. And in Item, tempThreshold is meant to be of type Double, not double. Try running the tests both ways. And finally, updated comments.